@@ -1,53 +1,89 @@
 import React from 'react';
 import { Button, Paper, Typography, List, ListItem, ListItemText, Grid } from '@mui/material';
 import { useNavigate, useParams } from 'react-router-dom';
-import data from './data';
+import { UseContext } from '../hook/useStatus';
+import axios from 'axios';
+
 function Voting() {
   const navigate = useNavigate();
+  const { setVc, setIsLogin, vc } = UseContext();
+  const [voteData, setVoteData] = React.useState({});
   const { id } = useParams();
-  
-  // Find the vote with the given id
-  const vote = data.find(vote => vote.id === Number(id));
+
+  React.useEffect(() => {
+    async function fetchData() {
+      const response = await axios.get(`/api/getPoll/${id}`);
+      setVoteData(response.data);
+      console.log(response.data);
+    }
+    fetchData();
+  }, []);
 
   const handleBack = () => {
     navigate('/voting-list');
   };
 
+  const handleVoting = async (poll_ID, choice) => {
+    const formData = new FormData();
+    formData.append('poll_ID', poll_ID);
+    formData.append('choice', choice);
+    formData.append('vc', vc);
+    const response = await axios.post('/api/voting', formData);
+    if(response == 'voted'){
+      console.log('voted');
+    }
+    else if(response == 'error'){
+      console.log('error');
+    }
+    else{
+      setVc(JSON.stringify(response.data));
+      navigate('/voting-list');
+    }
+
+    
+  };
+
   const handleLogout = () => {
+    setIsLogin(0);
+    setVc('');
     navigate('/');
   };
 
+  const dateTime = new Date(voteData.time);
+  const formattedTime = dateTime.toLocaleString();
+
   return (
     <Grid container direction="column" alignItems="center" spacing={2}>
-        <Grid item>
-            <Button variant="contained" color="secondary" onClick={handleLogout}>
-                Logout
-            </Button>
-            <Button variant="contained" color="primary" onClick={handleBack}>
-                Back to list
-            </Button>
-        </Grid>
       <Grid item>
-        <Paper sx={{ p: 2,width: '60vw' }}>
-          <Typography variant="h6">{vote.title}</Typography>
-          <Typography variant="body2">{vote.description}</Typography>
+        <Button variant="contained" color="secondary" onClick={handleLogout}>
+          Logout
+        </Button>
+        <Button variant="contained" color="primary" onClick={handleBack}>
+          Back to list
+        </Button>
+      </Grid>
+      <Grid item>
+        <Paper sx={{ p: 2, width: '60vw' }}>
+          <Typography variant="h6">{voteData.title}</Typography>
+          <Typography variant="body2">{voteData.description}</Typography>
           <List>
-            {vote.options.map((option, index) => (
-              <ListItem key={index}>
-                <ListItemText>
-                  {option.option}
-                  {vote.active ? (
-                    <Button variant="contained" color="primary">
-                      Vote
-                    </Button>
-                  ) : (
-                    ` - ${option.votes} votes`
-                  )}
-                </ListItemText>
-              </ListItem>
-            ))}
+            {voteData.options &&
+              voteData.options.map((option, index) => (
+                <ListItem key={index}>
+                  <ListItemText>
+                    {option}
+                    {!voteData.voted && voteData.active ? (
+                      <Button variant="contained" color="primary" onClick={()=>handleVoting(id,option)}>
+                        Vote
+                      </Button>
+                    ) : voteData.active ? null : (
+                      ` - ${option.votes} votes`
+                    )}
+                  </ListItemText>
+                </ListItem>
+              ))}
           </List>
-          <Typography variant="body2">End Time: {vote.endTime}</Typography>
+          <Typography variant="body2">End Time: {Date(voteData.time).toLocaleString()}</Typography>
         </Paper>
       </Grid>
     </Grid>
