@@ -5,6 +5,7 @@ const data = [
       description: '這是投票1的描述',
       options: [{ option: '選項1'}, { option: '選項2'}],
       active: true,
+      voted: false,
       endTime: '2023-12-31',
     },
     {
@@ -13,14 +14,16 @@ const data = [
       description: '這是投票2的描述',
       options: [{ option: '選項1', votes: 30 }, { option: '選項2', votes: 40 }, { option: '選項3', votes: 50 }],
       active: false,
+      voted: true,
       endTime: '2023-05-31',
     },
     {
         id: 3,
-        titl: '投票3',
+        title: '投票3',
         description:'這是投票3的描述',
         options: [{ option: '項1'}, { option: '選項2'}],
         active: true,
+        voted: true,
         endTime: '2023-12-31',
       },
       {
@@ -29,6 +32,7 @@ const data = [
         description: '這是投票4的描述',
         options: [{ option: '選項1', votes: 30 }, { option: '選項2', votes: 40 }, { option: '選項3', votes: 50 }],
         active: false,
+        voted: false,
         endTime: '2023-05-31',
       },
   ];
